fix: reload spriteImageList if it is not loaded
